Refactor database service and extract UI components
*   Split `DatabaseService` logic into specific repositories (`Model`, `Prompt`, `Task`, `Usage`) for better separation of concerns.
*   Extract `ModelSelectionSection` and `LibraryDialog` from `ControlPanelWidget` into separate files.
*   Extract `ModelEditDialog`, `ChannelEditDialog`, and `DiscoveryDialog` from `ModelsScreen` into dedicated widgets.
*   Move tag colors definition to `AppConstants`.
*   Simplify state management in `ControlPanelWidget` using `context.watch`.
*   Fix `initialValue` usage in `ImageDownloaderScreen` dropdown.
